python-whatsapp-transcription-bot/requirements.txt
azure-functions==1.17.0
openai==1.3.0
requests==2.31.0